[kbss-cvut/termit-ui#296] Refactor term export to CSV/Excel into separate classes to cleanup entity code.
