+++ src/test/java/cz/cvut/kbss/termit/service/export/CsvVocabularyExporterTest.java
@@ -14,15 +14,19 @@
  */
 package cz.cvut.kbss.termit.service.export;
 
+import cz.cvut.kbss.termit.dto.export.TabularTermExportUtils;
 import cz.cvut.kbss.termit.environment.Generator;
 import cz.cvut.kbss.termit.model.Term;
 import cz.cvut.kbss.termit.model.Vocabulary;
+import cz.cvut.kbss.termit.service.mapper.TermDtoMapper;
+import cz.cvut.kbss.termit.service.mapper.TermDtoMapperImpl;
 import cz.cvut.kbss.termit.service.repository.TermRepositoryService;
 import cz.cvut.kbss.termit.util.Constants;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
+import org.mockito.Spy;
 import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.core.io.Resource;
 import org.springframework.http.MediaType;
@@ -44,6 +48,9 @@ class CsvVocabularyExporterTest {
     @Mock
     private TermRepositoryService termService;
 
+    @Spy
+    private TermDtoMapper dtoMapper = new TermDtoMapperImpl();
+
     @InjectMocks
     private CsvVocabularyExporter sut;
 
@@ -55,7 +62,7 @@ class CsvVocabularyExporterTest {
         final Resource result = sut.exportGlossary(vocabulary, exportConfig());
         try (final BufferedReader reader = new BufferedReader(new InputStreamReader(result.getInputStream()))) {
             final String header = reader.readLine();
-            assertEquals(String.join(",", Term.EXPORT_COLUMNS), header);
+            assertEquals(String.join(",", TabularTermExportUtils.EXPORT_COLUMNS), header);
         }
     }
