+++ src/test/java/cz/cvut/kbss/termit/service/export/ExcelVocabularyExporterTest.java
@@ -1,25 +1,25 @@
 /**
- * TermIt
- * Copyright (C) 2019 Czech Technical University in Prague
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ * TermIt Copyright (C) 2019 Czech Technical University in Prague
+ * <p>
+ * This program is free software: you can redistribute it and/or modify it under the terms of the GNU General Public
+ * License as published by the Free Software Foundation, either version 3 of the License, or (at your option) any later
+ * version.
+ * <p>
+ * This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even the implied
+ * warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU General Public License for more
+ * details.
+ * <p>
+ * You should have received a copy of the GNU General Public License along with this program.  If not, see
+ * <https://www.gnu.org/licenses/>.
  */
 package cz.cvut.kbss.termit.service.export;
 
+import cz.cvut.kbss.termit.dto.export.TabularTermExportUtils;
 import cz.cvut.kbss.termit.environment.Generator;
 import cz.cvut.kbss.termit.model.Term;
 import cz.cvut.kbss.termit.model.Vocabulary;
+import cz.cvut.kbss.termit.service.mapper.TermDtoMapper;
+import cz.cvut.kbss.termit.service.mapper.TermDtoMapperImpl;
 import cz.cvut.kbss.termit.service.repository.TermRepositoryService;
 import cz.cvut.kbss.termit.util.Constants;
 import org.apache.poi.xssf.usermodel.XSSFRow;
@@ -29,6 +29,7 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
+import org.mockito.Spy;
 import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.core.io.Resource;
 import org.springframework.http.MediaType;
@@ -47,6 +48,9 @@ class ExcelVocabularyExporterTest {
 
     @Mock
     private TermRepositoryService termService;
+
+    @Spy
+    private TermDtoMapper dtoMapper = new TermDtoMapperImpl();
 
     @InjectMocks
     private ExcelVocabularyExporter sut;
@@ -78,8 +82,8 @@ class ExcelVocabularyExporterTest {
         assertNotNull(sheet);
         final XSSFRow row = sheet.getRow(0);
         assertNotNull(row);
-        for (int i = 0; i < Term.EXPORT_COLUMNS.size(); i++) {
-            assertEquals(Term.EXPORT_COLUMNS.get(i), row.getCell(i).getStringCellValue());
+        for (int i = 0; i < TabularTermExportUtils.EXPORT_COLUMNS.size(); i++) {
+            assertEquals(TabularTermExportUtils.EXPORT_COLUMNS.get(i), row.getCell(i).getStringCellValue());
         }
     }
